feat: organize elements.ts to a directory subdivided to keep more organized
